+++ tests/scripts/get-remediation-lapi-with-geoloc.php
@@ -6,6 +6,7 @@ use CrowdSec\LapiClient\Bouncer;
 use CrowdSec\RemediationEngine\CacheStorage\Memcached;
 use CrowdSec\RemediationEngine\CacheStorage\PhpFiles;
 use CrowdSec\RemediationEngine\CacheStorage\Redis;
+use CrowdSec\RemediationEngine\Constants;
 use CrowdSec\RemediationEngine\LapiRemediation;
 use CrowdSec\RemediationEngine\Logger\FileLog;
 
@@ -22,9 +23,10 @@ $bouncerKey = $argv[2] ?? false;
 $lapiUrl = $argv[3] ?? false;
 $streamMode = isset($argv[4]) ? (bool) $argv[4] : true;
 if (!$bouncerKey || !$lapiUrl) {
-    exit('Params <BOUNCER_KEY> and <LAPI_URL> are required' . \PHP_EOL
-         . 'Usage: php get-remediation-lapi.php <IP> <BOUNCER_KEY> <LAPI_URL> <STREAM_MODE>'
-         . \PHP_EOL);
+    exit('Params <BOUNCER_KEY> and <LAPI_URL> are required' . \PHP_EOL .
+         'Usage: php get-remediation-lapi.php <IP> <BOUNCER_KEY> <LAPI_URL> <STREAM_MODE>' .\PHP_EOL .
+         'Example: php get-remediation-lapi.php 172.0.0.24 c580ebdff45da6e01415ed0e9bc9c06b  https://crowdsec:8080 0' .
+         \PHP_EOL);
 }
 
 // Init  logger
@@ -52,9 +54,18 @@ $cacheRedisConfigs = [
     'redis_dsn' => 'redis://redis:6379',
 ];
 $redisCache = new Redis($cacheRedisConfigs, $logger);
-// Init LAPI remediation
+// Init LAPI remediation with geolocation
 $remediationConfigs = [
     'stream_mode' => $streamMode,
+    'geolocation' => [
+        'enabled' => true,
+        'cache_duration' => 120,
+        'type' => Constants::GEOLOCATION_TYPE_MAXMIND,
+        'maxmind' => [
+            'database_type' => Constants::MAXMIND_COUNTRY,
+            'database_path' => __DIR__ . '/../geolocation/GeoLite2-Country.mmdb',
+        ],
+    ],
 ];
 $remediationEngine = new LapiRemediation($remediationConfigs, $lapiClient, $phpFileCache, $logger);
 // Determine the remediation for the given IP